+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Modern URI 0.0.2.0
+
+* Added the `renderStr` and `renderStr'` functions for efficient rendering
+  to `String` and `ShowS`.
+
 ## Modern URI 0.0.1.0
 
 * Initial release.